+++ riscv/src/register/mcountinhibit.rs
@@ -1,66 +1,52 @@
-//! mcounteren register
+//! `mcountinhibit` register
 
 use crate::bits::{bf_extract, bf_insert};
 
-/// mcounteren register
+/// `mcountinhibit` register
 #[derive(Clone, Copy, Debug)]
-pub struct Mcounteren {
+pub struct Mcountinhibit {
     bits: usize,
 }
 
-impl Mcounteren {
-    /// Supervisor "cycle\[h\]" Enable
+impl Mcountinhibit {
+    /// Machine "cycle\[h\]" Disable
     #[inline]
     pub fn cy(&self) -> bool {
         bf_extract(self.bits, 0, 1) != 0
     }
 
-    /// Sets whether to enable the "cycle\[h\]" counter.
+    /// Sets whether to inhibit the "cycle\[h\]" counter.
     ///
-    /// Only updates the in-memory value, does not modify the `mcounteren` register.
+    /// Only updates the in-memory value, does not modify the `mcountinhibit` register.
     #[inline]
     pub fn set_cy(&mut self, cy: bool) {
         self.bits = bf_insert(self.bits, 0, 1, cy as usize);
     }
 
-    /// Supervisor "time\[h\]" Enable
-    #[inline]
-    pub fn tm(&self) -> bool {
-        bf_extract(self.bits, 1, 1) != 0
-    }
-
-    /// Sets whether to enable "time\[h\]".
-    ///
-    /// Only updates the in-memory value, does not modify the `mcounteren` register.
-    #[inline]
-    pub fn set_tm(&mut self, tm: bool) {
-        self.bits = bf_insert(self.bits, 1, 1, tm as usize);
-    }
-
-    /// Supervisor "instret\[h\]" Enable
+    /// Machine "instret\[h\]" Disable
     #[inline]
     pub fn ir(&self) -> bool {
         bf_extract(self.bits, 2, 1) != 0
     }
 
-    /// Sets whether to enable the "instret\[h\]" counter.
+    /// Sets whether to inhibit the "instret\[h\]" counter.
     ///
-    /// Only updates the in-memory value, does not modify the `mcounteren` register.
+    /// Only updates the in-memory value, does not modify the `mcountinhibit` register.
     #[inline]
     pub fn set_ir(&mut self, ir: bool) {
         self.bits = bf_insert(self.bits, 2, 1, ir as usize);
     }
 
-    /// Supervisor "hpm\[x\]" Enable (bits 3-31)
+    /// Machine "hpm\[x\]" Disable (bits 3-31)
     #[inline]
     pub fn hpm(&self, index: usize) -> bool {
         assert!((3..32).contains(&index));
         bf_extract(self.bits, index, 1) != 0
     }
 
-    /// Sets whether to enable the "hpm\[X\]" counter.
+    /// Sets whether to inhibit the "hpm\[X\]" counter.
     ///
-    /// Only updates the in-memory value, does not modify the `mcounteren` register.
+    /// Only updates the in-memory value, does not modify the `mcountinhibit` register.
     #[inline]
     pub fn set_hpm(&mut self, index: usize, hpm: bool) {
         assert!((3..32).contains(&index));
@@ -68,21 +54,17 @@ impl Mcounteren {
     }
 }
 
-read_csr_as!(Mcounteren, 0x306);
-write_csr_as!(Mcounteren, 0x306);
-set!(0x306);
-clear!(0x306);
+read_csr_as!(Mcountinhibit, 0x320);
+write_csr_as!(Mcountinhibit, 0x320);
+set!(0x320);
+clear!(0x320);
 
 set_clear_csr!(
-/// Supervisor cycle Enable
+/// Machine cycle Disable
     , set_cy, clear_cy, 1 << 0);
 
 set_clear_csr!(
-/// Supervisor time Enable
-    , set_tm, clear_tm, 1 << 1);
-
-set_clear_csr!(
-/// Supervisor instret Enable
+/// Machine instret Disable
     , set_ir, clear_ir, 1 << 2);
 
 #[inline]
@@ -102,8 +84,8 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_mcounteren() {
-        let mut m = Mcounteren { bits: 0 };
+    fn test_mcountinhibit() {
+        let mut m = Mcountinhibit { bits: 0 };
 
         assert!(!m.cy());
 
@@ -112,14 +94,6 @@ mod tests {
 
         m.set_cy(false);
         assert!(!m.cy());
-
-        assert!(!m.tm());
-
-        m.set_tm(true);
-        assert!(m.tm());
-
-        m.set_tm(false);
-        assert!(!m.tm());
 
         assert!(!m.ir());